+++ src/utils/database/initialize_database.py
@@ -11,11 +11,12 @@ def drop_tables(connection):
 
     cursor = connection.cursor()
 
-    cursor.execute("""
+    cursor.executescript("""
+        BEGIN;
         DROP TABLE IF EXISTS Users;
+        DROP TABLE IF EXISTS Cards;
+        COMMIT;
     """)
-
-    connection.commit()
 
 
 def create_tables(connection):
@@ -28,15 +29,38 @@ def create_tables(connection):
 
     cursor = connection.cursor()
 
-    cursor.execute("""
+    cursor.executescript("""
+        BEGIN;
         CREATE TABLE Users (
             id INTEGER PRIMARY KEY,
             username TEXT,
             password TEXT
         );
+        CREATE TABLE Cards (
+            id INTEGER PRIMARY KEY,
+            name TEXT,
+            released_at TEXT,
+            layout TEXT,
+            mana_cost TEXT,
+            cmc REAL,
+            colors JSON,
+            color_identity JSON,
+            type_line TEXT,
+            oracle_text TEXT,
+            keywords JSON,
+            card_faces JSON,
+            all_parts JSON,
+            power TEXT,
+            toughness TEXT,
+            image_uris JSON,
+            set_code TEXT,
+            set_name TEXT,
+            rarity TEXT,
+            flavor_text TEXT,
+            prices JSON
+        );
+        COMMIT;
     """)
-
-    connection.commit()
 
 
 def initialize_database():